requirements.txt
Flask==2.3.2
keras==2.13.1
numpy==1.24.3
Pillow==10.0.0
Pillow==10.0.0
scikit_learn==1.3.0
tensorflow==2.13.0
tensorflow_intel==2.13.0
gunicorn==20.0.4
Werkzeug==0.16.0
Jinja2==2.10.1
itsdangerous==1.1.0
MarkupSafe==1.1.1